Changed sentiment score from float to double data type.

+++ portfolio/src/main/java/com/google/sps/data/Comment.java
@@ -7,10 +7,10 @@ public final class Comment {
   private final long id;
   private final String text;
   private final String name;
-  private final float score;
+  private final double score;
   private final Date date;
 
-  public Comment(long id, String text, String name, float score, long time) {
+  public Comment(long id, String text, String name, double score, long time) {
     this.id = id;
     this.name = name;
     this.text = text;
